chore: rename/replace method firebase global new;
