+++ src/lib/firebase/firebase.actions.ts
@@ -12,7 +12,7 @@ import type { FIREBASE_odds } from '@betarena/scores-lib/types/firebase.js';
  * @param {object} fixtures_arr_filter 
  * @returns 
  */
-export async function getOdds
+export async function getOdds_1
 (
 	fixtures_arr_filter: {
 		date: Date;
@@ -77,6 +77,61 @@ export async function getOdds
 	}
 
 	return fixtureOddsMap;
+}
+
+/**
+ * @summary [MAIN]
+ * @description obtains array of 
+ * target FIXTURE odds;
+ * @param {string} fixture_time 
+ * @param {number} fixture_id 
+ * @returns 
+ */
+export async function getOdds_2
+(
+	fixture_time: string,
+	fixture_id: number
+): Promise < FIREBASE_odds[] > 
+{
+	const sportbook_array: FIREBASE_odds[] = [];
+
+	const year_: string = new Date(fixture_time).getFullYear().toString();
+	const month_: number = new Date(fixture_time).getMonth();
+	let new_month_ = (month_ + 1).toString();
+	new_month_ = `0${new_month_}`.slice(-2);
+	let day_ = new Date(fixture_time).getDate().toString();
+	day_ = `0${day_}`.slice(-2);
+
+	await get
+  (
+		child
+    (
+			ref
+      (
+        db_real
+      ),
+			`odds/${year_}/${new_month_}/${day_}/${fixture_id}`
+		)
+	)
+  .then
+  (
+    (
+      snapshot
+    ) => 
+    {
+      if (snapshot.exists()) 
+      {
+        const data: [string, FIREBASE_odds][] =	Object.entries(snapshot.val());
+        for (const sportbook of data) 
+        {
+          sportbook[1].sportbook = sportbook[0].toString();
+          sportbook_array.push(sportbook[1]);
+        }
+      }
+	  }
+  );
+
+	return sportbook_array;
 }
 
 /**
